break out models, DTOs, contexts, etc. into their own files for maintainability
