+++ backend/DTOs/RecipeDTO.cs
@@ -1,0 +1,4 @@
+public class RecipeDTO
+{
+  public string? Name { get; set; }
+}
