+++ backend/Data/RecipeDbContext.cs
@@ -1,0 +1,7 @@
+using Microsoft.EntityFrameworkCore;
+
+public class RecipeContext : DbContext
+{
+  public RecipeContext(DbContextOptions<RecipeContext> options) : base(options) { }
+  public DbSet<Recipe> Recipes { get; set; }
+}
