+++ backend/Models/Recipe.cs
@@ -1,0 +1,5 @@
+public class Recipe
+{
+  public int Id { get; set; }
+  public string? Name { get; set; }
+}
